requirements.txt
# Core AI and ML dependencies
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Explainability and Interpretability
shap>=0.42.0
lime>=0.2.0.1
interpret>=0.4.0

# Web Interface
streamlit>=1.28.0
gradio>=3.40.0

# Data Processing and Management
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0

# Security and Configuration
python-dotenv>=1.0.0
cryptography>=41.0.0
bcrypt>=4.0.0

# Monitoring and Logging
loguru>=0.7.0
prometheus-client>=0.17.0
mlflow>=2.6.0

# Cloud and Containerization
boto3>=1.28.0
kubernetes>=27.0.0
docker>=6.1.0

# Financial Data
yfinance>=0.2.18
alpha-vantage>=2.3.1
quandl>=3.7.0

# Additional Utilities
requests>=2.31.0
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0
celery>=5.3.0
flower>=2.0.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 